+++ Editor/SettingsAnalyzers/StaticBatchingAndHybridPackage.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using UnityEditor;
-
 #if UNITY_2019_3_OR_NEWER
 using PackageInfo = UnityEditor.PackageManager.PackageInfo;
 #endif